requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
numpy>=1.24.0
pillow>=9.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
opencv-python>=4.8.0
